refactor(print): share per-serving nutrition math; dedup Calories row
- Extract getQuantity + safeNutrientMap into src/util/nutrition.ts so the print
  layout and the on-screen NutritionData card share the per-serving math and
  missing-key safety (each keeps its own labels/markup)
- Route the print Calories row through the same row() helper as the other
  nutrients instead of an inline copy of Math.round(num/servings)

Pure refactor — output unchanged (verified: print rows and NutritionData tests
identical), addresses review findings #1 (partial) and #2.
